% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guescini-package.R
\docType{package}
\name{guescini-package}
\alias{guescini}
\alias{guescini-package}
\title{guescini: Real-Time PCR Data Sets by Guescini et al. (2008)}
\description{
Real-time quantitative polymerase chain reaction (qPCR) data by Guescini et al. (2008) \doi{10.1186/1471-2105-9-326} in tidy format. This package provides two data sets where the amplification efficiency has been modulated: either by alterating the amplification mix concentration, or by increasing the concentration of IgG, a PCR inhibitor. Original raw data files: \url{https://static-content.springer.com/esm/art\%3A10.1186\%2F1471-2105-9-326/MediaObjects/12859_2008_2311_MOESM1_ESM.xls} and \url{https://static-content.springer.com/esm/art\%3A10.1186\%2F1471-2105-9-326/MediaObjects/12859_2008_2311_MOESM5_ESM.xls}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/guescini}
  \item \url{http://rmagno.eu/guescini/}
  \item Report bugs at \url{https://github.com/ramiromagno/guescini/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
